+++ pkg/scheduler/plugins/minruntime/minruntime_suite_test.go
@@ -1,0 +1,16 @@
+// Copyright 2025 NVIDIA CORPORATION
+// SPDX-License-Identifier: Apache-2.0
+
+package minruntime
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestMinruntime(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Minruntime Plugin Suite")
+}