+++ test/e2e/modules/configurations/feature_flags/scheduler_config_map.go
@@ -50,6 +50,7 @@ var defaultKaiSchedulerPlugins = []tier{
 			{Name: "taskorder"},
 			{Name: "nominatednode"},
 			{Name: "dynamicresources"},
+			{Name: "minruntime"},
 		},
 	},
 }