+++ CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Removed
 - Removed `runai-job-id` and `runai/job-id` annotations from pods and podgroups
 
+### Added
+ - Added [minruntime](docs/plugins/minruntime.md) plugin, allowing PodGroups to run for a configurable amount of time without being reclaimed/preempted.
+
+
 ## [v0.5.1] - 2025-05-20
 
 ### Added